requirements.txt
simpleitk>=2.1
pydicom
tqdm
pyyaml
pandas
numpy
pyarrow